refactor: extract persisted settings into an AppSettings store (#452)
Move the persisted user settings (audio source, content region, language,
lyrics-anim direction, notification buttons, player gradient, canvas toggle)
plus loadPreferences/effectiveRegion and the setters — including their side
effects (notification push to the service, locale change) — out of
PlaybackViewModel into a process-scoped AppSettings store (like SessionHolder/
Navigator). PlaybackViewModel reads AppSettings in the stream-resolution path;
the UI reads/writes it directly. canvasUrl (the current track's video URL,
playback state, not persisted) stays on PlaybackViewModel; its thin
setCanvasEnabled wraps AppSettings.setCanvasEnabled to also clear the URL.
Service wiring reads notification buttons straight from prefs so a headphone
cold-start still gets saved values before AppSettings.load runs.

Slims PlaybackViewModel ~150 lines. Behavior-preserving; device-verified:
a fresh track resolves + plays (the hot-path audio-source/region reads),
settings render with their saved values, playback advances.

Closes #452
